% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogit_perm.R
\name{confint.clogit_perm}
\alias{confint.clogit_perm}
\title{Confidence interval for CLogit}
\usage{
\method{confint}{clogit_perm}(object, parm, level = 0.95, which. = "coef", sigma_perm = FALSE, ...)
}
\arguments{
\item{object}{An object of class \link{clogit_perm}}

\item{parm}{Integer vector. Indicates which parameters to include in the
output.}

\item{level}{Numeric. Level (see \link[stats:confint]{stats::confint}).}

\item{which.}{When \code{"coef"}, it will generate the confidence interval for the
parameter estimates, otherwise it generates the confidence interval of the reference
distribution.}

\item{sigma_perm}{Logical scalar. When \code{TRUE}, uses the permutation based
standard errors for computing the CI (only used when \code{which. = "coef"})}

\item{...}{Ignored.}
}
\value{
A two-column matrix with one row per parameter, giving the lower and upper
bounds of the interval. Column names record the percentiles used.
}
\description{
Confidence interval for CLogit
}
\seealso{
clogit_perm
}
