% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure_dyn}
\alias{exposure_dyn}
\title{Dynamic Exposure}
\usage{
exposure_dyn(id_indiv, id_events, actions, offset = 1)
}
\arguments{
\item{id_indiv}{Integer vector. Ids of individuals}

\item{id_events}{Integer vector. Ids of the events (must have a temporal
mapping).}

\item{actions}{Integer vector. Any value greater than one is considered
to be an action (true).}

\item{offset}{Integer. Offset in terms of lead}
}
\value{
A matrix with the following columns:
\itemize{
\item cumsum Integer vector. Cumulative sum of the actions vector.
\item exposure_i Integer vector. Indirect exposure.
\item exposure_d Integer vector. Direct exposure.
\item exposure_i_cum Integer vector. Cumulative indirect exposure.
\item exposure_d_cum Integer vector. Cumulative direct exposure.
}
}
\description{
Computes dynamic exposure in a longitudinal dataset without double counting
individuals.
}
\examples{
dat <- data.frame(
  event = c(1,1,1, 2,2,2, 3,3,3,3, 4,4,4,4),
  id    = c(1,2,3, 1,2,4, 2,3,4,5, 5,1,6,2),
  fired = c(1,1,1, 0,1,1, 0,1,0,1, 1,0,1,1)
)

counts <- with(dat, exposure_dyn(
  id_indiv  = id,
  id_events = event,
  actions   = fired,
  offset    = 0
))

cbind(dat, counts)
}
