% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.clogit_perm}
\alias{plot.clogit_perm}
\title{Plot method for \link{clogit_perm} objects}
\usage{
\method{plot}{clogit_perm}(
  x,
  y = NULL,
  level = 0.95,
  col = 1:length(stats::coef(x)),
  args_points = list(pch = 19, cex = 1.5),
  args_arrows = list(lwd = 2, code = 3, angle = 90, length = 0.1),
  labels = NULL,
  odds = FALSE,
  which. = "coef",
  ...
)
}
\arguments{
\item{x}{An object of class \link{clogit_perm}.}

\item{y}{Ignored}

\item{level}{Passed to \code{\link[stats:confint]{stats::confint()}}.}

\item{col}{Vector of colors for the model terms.}

\item{args_points}{List of arguments passed to \code{\link[graphics:points]{graphics::points()}}}

\item{args_arrows}{List of arguments passed to \code{\link[graphics:arrows]{graphics::arrows()}}}

\item{labels}{a named list with alternative labels for the model terms.}

\item{odds}{See \link{confint.clogit_perm}.}

\item{which.}{See \link{confint.clogit_perm}.}

\item{...}{Ignored}
}
\value{
\code{NULL}, invisibly. Called for its side effect of drawing a forest plot of
the coefficients (or odds ratios) with their permutation-based confidence
intervals.
}
\description{
Plot method for \link{clogit_perm} objects
}
