% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_events}
\alias{sim_events}
\alias{sim_events2}
\title{Simulate Police Force Events}
\usage{
sim_events(
  nevents,
  nofficers,
  min_per_event = 1,
  max_per_event = 5,
  min_year = 0,
  max_year = 10,
  min_rate = 5,
  max_rate = 5,
  par_officer_female = -0.5,
  par_officer_years = -0.5,
  par_exposure_event = 0,
  par_exposure_prev = 0.5,
  par_event_violence = 1,
  par_officer_fe = 1,
  nsims = 1,
  seed = sample.int(.Machine$integer.max, 1)
)

sim_events2(
  event_id,
  officer_id,
  officer_female,
  officer_years,
  officer_fe = rep(0, length(event_id)),
  officer_rate = rep(1, length(event_id)),
  par_officer_female = -0.5,
  par_officer_years = -0.5,
  par_exposure_event = 0.5,
  par_exposure_prev = 0.5,
  par_event_violence = 1,
  par_officer_fe = 1,
  nsims = 1,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{nevents, nofficers}{Integers. Number of events and officers to simulate.}

\item{min_per_event, max_per_event}{Integers. Lower and upper bounds for the
number of officers in the event.}

\item{min_year, max_year}{Integers. Lower and upper bounds for the number
of years of experience of the officers.}

\item{min_rate, max_rate}{Doubles. Lower and upper bounds for the reaction
rates (see details).}

\item{par_officer_female, par_officer_years, par_exposure_event, par_exposure_prev, par_event_violence, par_officer_fe}{Doubles. Parameters (coefficients) for the logistic probabilities.}

\item{nsims}{Integer. When greater than 1, the program will simulate multiple
datasets, using the same set of parameters and individual's features, and
append the corresponding results to the end of data frame.}

\item{seed}{Integer. Seed for the pseudo-number generation.}

\item{event_id, officer_id}{Integer vectors. Values for the incident and
officer id.}

\item{officer_female, officer_years}{Logical and integer vectors, respectively. Features
of the officers.}

\item{officer_fe, officer_rate}{Double vectors, more features of the officers.}
}
\value{
A data frame with the following columns
\itemize{
\item \code{officerid} Id of the police officer
\item \code{female} 1 if it is female
\item \code{years} years of experience
\item \code{fixed_effect} officers' propensity to point the gun
\item \code{incidentid} Incident id
\item \code{violence_level} Violence level of the event
\item \code{response_time} Time the officer took to respond
\item \code{first} Whether the officer was the first to act or not
\item \code{exposed} Whether the officer was exposed in the previous event
\item \verb{pointed000001 ... pointed[nsims]} Integer vectors indicating whether the
officer pointed the gun or not.
}

Each row represents one report per officer involved in the event.
}
\description{
This function generates data similar to that featured in the paper. Events
are drawn at random, as the number of officers per event. The outcome variable,
whether the officer points his gun or not, is drawn sequentially as a poisson
process.
}
\details{
The simulation process goes as follow:
\enumerate{
\item The officers are simulated.

\deqn{Female ~ Bernoulli(0.5)}

\deqn{Action rate ~ Unif(min_rate, max_rate)}

\deqn{Years of experience ~ Discrete Unif[min_years, max_year]}
\item Events are simulated, each event has

\deqn{nofficers ~ Discrete Unif[min_per_event, max_per_event]}

Once the event is done, a sequence of reaction is given by each officers'
action rate (Poisson process). Whether an officer points or not is set by
a logistic model

\deqn{
   point ~ logis(female + years of experience + has any pointed? + previous exposure)
   }

The corresponding parameters are as specified by the user. Events are simulated
one at a time.
}

In the case of \code{sim_events2}, the user can pass predefined events and
officers and use those to simulate each officers' reactions.
}
\examples{
x <- sim_events(1000, 400)

x <- sim_events(
  20000,200,
  par_officer_female = -.5,
  par_officer_years = -.5,
  par_exposure_event = -.5,
  par_event_violence = 1,
  par_exposure_prev = .25,
  par_officer_fe = 1,
  seed = 445
)

# Full model knowing latent variables
ans <- glm(
  pointed000001 ~ -1+female + years + exposed + I(-first) + fixed_effect +
  violence_level, data = x, family = binomial()
)

summary(ans)
}
