% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute}
\alias{permute}
\title{Random permutation of the data as a function of \code{find_candidates}}
\usage{
permute(candidates)
}
\arguments{
\item{candidates}{An integer list as that resulting from \code{\link[=find_candidates]{find_candidates()}}.}
}
\value{
An integer vector (indexed from 0) with the permuted version of the data.
}
\description{
Random permutation of the data as a function of \code{find_candidates}
}
