% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.clogit_perm}
\alias{print.clogit_perm}
\title{Print method for clogit_perm}
\usage{
\method{print}{clogit_perm}(x, odds = TRUE, labels = NULL, out = "ascii", ...)
}
\arguments{
\item{x}{An object of class \link{clogit_perm}.}

\item{odds}{Logical scalar. When TRUE it prints the odds ratios.}

\item{labels}{Named vector. Changes the labels of the model.}

\item{out}{Character scalar. When \code{"ascii"} it will print for screen, otherwise,
it will print for LaTeX.}

\item{...}{Ignored.}
}
\description{
Print method for clogit_perm
}
